conversations: extract DraftCards.svelte from ConversationsPanel (#1087, increment 2)
Second of the ConversationsPanel split — lift the draft-card / result-line
rendering into conversations/DraftCards.svelte. One component renders every
anchored card for a given anchor: `<DraftCards {tab} index={i} />` for a message
index, `index={null}` for the orphan stack (cards anchored past the current
message list). This collapses the two near-identical ~70-line interleave blocks
(per-message + orphans) in the panel into two one-line component uses.

DraftCards owns the 10 card/result snippets, all the approve/discard handlers,
the compute-card forwarders, openFiled*/toggleDraftPath, its own expandedDraftIds,
and the `pick`/`pickResults` anchor filters — reaching the conversations/editor
stores via their getters, so it only needs {tab, index} as props. The card-body
CSS (draft-paths / source-list / property / claims / filed-line) moves with it.

The message rendering, citations (citeState/handleCite), streaming, ask-user
card, and scroll stay in the panel (increment 3 will extract those as MessageList).

ConversationsPanel.svelte: 1695 → 848 LOC. pnpm lint + full pnpm test (4043) green.
