+++ src/renderer/lib/components/ConversationsPanel.svelte
@@ -17,49 +17,14 @@
     isEffort,
     type Effort,
   } from '../../../shared/tools/effort';
-  import type { ConversationDraft } from '../../../shared/conversation-drafts';
-  import type {
-    ConversationSourceDraft,
-    SourceIngestOutcome,
-  } from '../../../shared/conversation-source-drafts';
-  import type {
-    ConversationPropertyDraft,
-    PropertyUpdateOutcome,
-  } from '../../../shared/conversation-property-drafts';
-  import type {
-    ConversationSourcePropertyDraft,
-    SourcePropertyOutcome,
-  } from '../../../shared/conversation-source-property-drafts';
-  import type {
-    ConversationClaimsDraft,
-    ClaimsOutcome,
-  } from '../../../shared/conversation-claims-drafts';
-  import type {
-    ConversationComputeDraft,
-  } from '../../../shared/conversation-compute-drafts';
   import type { ConversationMessage, Citation } from '../../../shared/types';
   import { insertCitationMarker } from '../conversations/cite-from-conversation';
   import { type CiteStatus } from '../conversations/citations';
   import MessageCitations from './MessageCitations.svelte';
-  import DraftCard from './DraftCard.svelte';
-  import ComputeDraftCard from './ComputeDraftCard.svelte';
-  import RefactorDraftCard from './RefactorDraftCard.svelte';
-  import ReorgDraftCard from './ReorgDraftCard.svelte';
-  import DeleteDraftCard from './DeleteDraftCard.svelte';
-  import NoteBodyDraftCard from './NoteBodyDraftCard.svelte';
   import Composer from './conversations/Composer.svelte';
-  import type { ConversationRefactorDraft, ConversationReorgDraft, ConversationDeleteDraft } from '../../../shared/conversation-refactor-drafts';
-  import type { ConversationNoteBodyDraft } from '../../../shared/conversation-note-body-drafts';
-  import {
-    tabTitle,
-    formatPropertyValue,
-    sourceLabel,
-    basename,
-    sourceKindLabel,
-  } from '../conversations/conversation-display';
-  import {
-    formatTurnCost,
-  } from '../conversations/conversation-cost';
+  import DraftCards from './conversations/DraftCards.svelte';
+  import { tabTitle } from '../conversations/conversation-display';
+  import { formatTurnCost } from '../conversations/conversation-cost';
 
   // Lightweight markdown-it for assistant message rendering. Mirrors the
   // configuration in the legacy ConversationDialog so prose renders the
@@ -100,7 +65,6 @@
 
   let composer = $state<{ focus: () => void }>();
   let scrollEl = $state<HTMLDivElement>();
-  let expandedDraftIds = $state<Set<string>>(new Set());
   let resizing = $state(false);
   // Project-default model — used to label the "Default" option so the
   // user can see which concrete model "default" resolves to.
@@ -305,230 +269,10 @@
     await store.answerQuestion(tabId, text);
   }
 
-  function toggleDraftPath(key: string) {
-    const next = new Set(expandedDraftIds);
-    if (next.has(key)) next.delete(key);
-    else next.add(key);
-    expandedDraftIds = next;
-  }
 
-  async function handleApprove(tabId: string, draft: ConversationDraft) {
-    try {
-      const { filedPaths } = await store.approveDraft(tabId, draft);
-      // Open the first filed note so the user lands on what they just approved.
-      if (filedPaths.length > 0) void editor.openFile(filedPaths[0]!);
-    } catch (e) {
-      console.error('[conv-panel] approve failed:', e);
-    }
-  }
-
-  function handleDiscard(tabId: string, draftId: string) {
-    store.discardDraft(tabId, draftId);
-  }
-
-  async function handleApproveRefactor(tabId: string, draft: ConversationRefactorDraft) {
-    try {
-      await store.approveRefactorDraft(tabId, draft);
-      // Land the user on the note at its new path.
-      void editor.openFile(draft.toPath);
-    } catch (e) {
-      console.error('[conv-panel] approve refactor failed:', e);
-    }
-  }
-
-  function handleDiscardRefactor(tabId: string, draftId: string) {
-    store.discardRefactorDraft(tabId, draftId);
-  }
-
-  async function handleApproveReorg(
-    tabId: string, draft: ConversationReorgDraft, selected: Array<{ fromPath: string; toPath: string }>,
-  ) {
-    try {
-      await store.approveReorgDraft(tabId, draft, selected);
-      // Land on the first moved note at its new path.
-      if (selected.length > 0) void editor.openFile(selected[0]!.toPath);
-    } catch (e) {
-      console.error('[conv-panel] approve reorg failed:', e);
-    }
-  }
-
-  function handleDiscardReorg(tabId: string, draftId: string) {
-    store.discardReorgDraft(tabId, draftId);
-  }
-
-  async function handleApproveDelete(
-    tabId: string, draft: ConversationDeleteDraft, selected: string[],
-  ) {
-    try {
-      await store.approveDeleteDraft(tabId, draft, selected);
-    } catch (e) {
-      console.error('[conv-panel] approve delete failed:', e);
-    }
-  }
-
-  function handleDiscardDelete(tabId: string, draftId: string) {
-    store.discardDeleteDraft(tabId, draftId);
-  }
-
-  async function handleApproveNoteBody(tabId: string, draft: ConversationNoteBodyDraft) {
-    try {
-      await store.approveNoteBodyDraft(tabId, draft);
-    } catch (e) {
-      console.error('[conv-panel] approve note-body rewrite failed:', e);
-    }
-  }
-
-  function handleDiscardNoteBody(tabId: string, draftId: string) {
-    store.discardNoteBodyDraft(tabId, draftId);
-  }
-
-  async function handleApproveSource(tabId: string, draft: ConversationSourceDraft) {
-    try {
-      await store.approveSourceDraft(tabId, draft);
-    } catch (e) {
-      console.error('[conv-panel] approve source failed:', e);
-    }
-  }
-
-  function handleDiscardSource(tabId: string, draftId: string) {
-    store.discardSourceDraft(tabId, draftId);
-  }
-
-  async function handleApproveProperty(tabId: string, draft: ConversationPropertyDraft) {
-    try {
-      await store.approvePropertyDraft(tabId, draft);
-    } catch (e) {
-      console.error('[conv-panel] approve property failed:', e);
-    }
-  }
-
-  function handleDiscardProperty(tabId: string, draftId: string) {
-    store.discardPropertyDraft(tabId, draftId);
-  }
-
-  async function handleApproveSourceProperty(tabId: string, draft: ConversationSourcePropertyDraft) {
-    try {
-      await store.approveSourcePropertyDraft(tabId, draft);
-    } catch (e) {
-      console.error('[conv-panel] approve source property failed:', e);
-    }
-  }
-
-  function handleDiscardSourceProperty(tabId: string, draftId: string) {
-    store.discardSourcePropertyDraft(tabId, draftId);
-  }
-
-  async function handleApproveClaims(tabId: string, draft: ConversationClaimsDraft) {
-    try {
-      await store.approveClaimsDraft(tabId, draft);
-    } catch (e) {
-      console.error('[conv-panel] approve claims failed:', e);
-    }
-  }
-
-  function handleDiscardClaims(tabId: string, draftId: string) {
-    store.discardClaimsDraft(tabId, draftId);
-  }
-
-  // ── propose_compute card actions (#245) ───────────────────────────
-  //
-  // The edit buffer / edit mode / risky-ack now live inside ComputeDraftCard
-  // (per-card view state); the panel only forwards the three store actions and
-  // the run state. `edited` is undefined when the cell was never edited.
-  function onRunCompute(draft: ConversationComputeDraft, edited: string | undefined): void {
-    const tab = store.activeTab;
-    if (!tab) return;
-    void store.runComputeDraft(tab.id, draft, edited);
-  }
-
-  function onInsertCompute(draft: ConversationComputeDraft, edited: string | undefined): void {
-    const tab = store.activeTab;
-    if (!tab) return;
-    void store.insertComputeDraft(tab.id, draft, edited);
-  }
-
-  function onDiscardCompute(draft: ConversationComputeDraft): void {
-    const tab = store.activeTab;
-    if (!tab) return;
-    store.discardComputeDraft(tab.id, draft.draftId);
-  }
-
-  function openInsertedNote(path: string): void {
-    void editor.openFile(path);
-  }
-
-  /** Render a frontmatter value for inline display on the review card.
-   *  Strings render bare; everything else falls back to compact JSON
-   *  so the user can eyeball arrays/objects without scrolling. Null
-   *  is shown as a deletion marker. */
-  function openFiledNote(relativePath: string) {
-    void editor.openFile(relativePath);
-  }
-
-  function openFiledSource(sourceId: string) {
-    editor.openSource(sourceId);
-  }
-
-  // ── Card anchoring helpers ───────────────────────────────────────────
-  // Each draft / sourceDraft / sourceDraftResult carries an
-  // `afterMessageIndex` captured when it arrived (the slot the streaming
-  // assistant message will occupy after end-of-send reload). To keep
-  // the visual conversation chronological, we walk messages and render
-  // each card right after the message it's anchored to. Cards whose
-  // anchor index is beyond the current message list (in-flight turn,
-  // canceled turn, etc.) render as orphans at the bottom — same
-  // position as the old "everything at panel bottom" behavior but only
-  // for the narrow window before reload lands.
+  // The active-tab shape, for the cite handlers above. Draft/result card
+  // anchoring (per-message vs orphan) now lives in DraftCards.svelte (#1087).
   type TabT = NonNullable<typeof store.activeTab>;
-  // Draft/result cards are anchored to a message index. Four generic filters
-  // capture the only shapes the panel renders by (#980 H5); the per-kind
-  // helpers below are thin, typed adapters so the template stays unchanged.
-  // `*At` → cards for message `i`; `orphan*` → cards anchored past the current
-  // message list (in-flight turn or post-cancel), rendered at the bottom.
-  function atIndex<T extends { afterMessageIndex: number }>(arr: T[], i: number): T[] {
-    return arr.filter((d) => d.afterMessageIndex === i);
-  }
-  function orphansOf<T extends { afterMessageIndex: number }>(tab: TabT, arr: T[]): T[] {
-    return arr.filter((d) => d.afterMessageIndex >= tab.conversation.messages.length);
-  }
-  function resultsAtIndex<E extends { afterMessageIndex: number }>(rec: Record<string, E>, i: number): [string, E][] {
-    return Object.entries(rec).filter(([, entry]) => entry.afterMessageIndex === i);
-  }
-  function orphanResultsOf<E extends { afterMessageIndex: number }>(tab: TabT, rec: Record<string, E>): [string, E][] {
-    return Object.entries(rec).filter(([, entry]) => entry.afterMessageIndex >= tab.conversation.messages.length);
-  }
-
-  function draftsAt(tab: TabT, i: number) { return atIndex(tab.drafts, i); }
-  function sourceDraftsAt(tab: TabT, i: number) { return atIndex(tab.sourceDrafts, i); }
-  function sourceResultsAt(tab: TabT, i: number) { return resultsAtIndex(tab.sourceDraftResults, i); }
-  function noteResultsAt(tab: TabT, i: number) { return resultsAtIndex(tab.noteDraftResults, i); }
-  function propertyDraftsAt(tab: TabT, i: number) { return atIndex(tab.propertyDrafts, i); }
-  function propertyResultsAt(tab: TabT, i: number) { return resultsAtIndex(tab.propertyDraftResults, i); }
-  function sourcePropertyDraftsAt(tab: TabT, i: number) { return atIndex(tab.sourcePropertyDrafts, i); }
-  function sourcePropertyResultsAt(tab: TabT, i: number) { return resultsAtIndex(tab.sourcePropertyDraftResults, i); }
-  function claimsDraftsAt(tab: TabT, i: number) { return atIndex(tab.claimsDrafts, i); }
-  function claimsResultsAt(tab: TabT, i: number) { return resultsAtIndex(tab.claimsDraftResults, i); }
-  function computeDraftsAt(tab: TabT, i: number) { return atIndex(tab.computeDrafts, i); }
-  function refactorDraftsAt(tab: TabT, i: number) { return atIndex(tab.refactorDrafts, i); }
-  function reorgDraftsAt(tab: TabT, i: number) { return atIndex(tab.reorgDrafts, i); }
-  function deleteDraftsAt(tab: TabT, i: number) { return atIndex(tab.deleteDrafts, i); }
-  function noteBodyDraftsAt(tab: TabT, i: number) { return atIndex(tab.noteBodyDrafts, i); }
-
-  function orphanDrafts(tab: TabT) { return orphansOf(tab, tab.drafts); }
-  function orphanSourceDrafts(tab: TabT) { return orphansOf(tab, tab.sourceDrafts); }
-  function orphanSourceResults(tab: TabT) { return orphanResultsOf(tab, tab.sourceDraftResults); }
-  function orphanNoteResults(tab: TabT) { return orphanResultsOf(tab, tab.noteDraftResults); }
-  function orphanPropertyDrafts(tab: TabT) { return orphansOf(tab, tab.propertyDrafts); }
-  function orphanPropertyResults(tab: TabT) { return orphanResultsOf(tab, tab.propertyDraftResults); }
-  function orphanSourcePropertyDrafts(tab: TabT) { return orphansOf(tab, tab.sourcePropertyDrafts); }
-  function orphanSourcePropertyResults(tab: TabT) { return orphanResultsOf(tab, tab.sourcePropertyDraftResults); }
-  function orphanClaimsDrafts(tab: TabT) { return orphansOf(tab, tab.claimsDrafts); }
-  function orphanClaimsResults(tab: TabT) { return orphanResultsOf(tab, tab.claimsDraftResults); }
-  function orphanComputeDrafts(tab: TabT) { return orphansOf(tab, tab.computeDrafts); }
-  function orphanRefactorDrafts(tab: TabT) { return orphansOf(tab, tab.refactorDrafts); }
-  function orphanReorgDrafts(tab: TabT) { return orphansOf(tab, tab.reorgDrafts); }
-  function orphanDeleteDrafts(tab: TabT) { return orphansOf(tab, tab.deleteDrafts); }
-  function orphanNoteBodyDrafts(tab: TabT) { return orphansOf(tab, tab.noteBodyDrafts); }
 </script>
 
 {#if store.visible}
@@ -662,249 +406,6 @@
           {/if}
         {/snippet}
 
-        {#snippet noteDraftCard(draft: ConversationDraft)}
-          <DraftCard
-            headline={`${draft.payloads.length} note${draft.payloads.length === 1 ? '' : 's'}`}
-            note={draft.note}
-            approveLabel="Approve & file"
-            onApprove={() => handleApprove(tab.id, draft)}
-            onDiscard={() => handleDiscard(tab.id, draft.draftId)}
-          >
-            <ul class="draft-paths">
-              {#each draft.payloads as p}
-                {@const key = draft.draftId + ':' + p.relativePath}
-                <li>
-                  <button type="button" class="draft-path-btn" onclick={() => toggleDraftPath(key)}>
-                    <span class="draft-path">{p.relativePath}</span>
-                    <span class="draft-toggle">{expandedDraftIds.has(key) ? '▾' : '▸'}</span>
-                  </button>
-                  {#if expandedDraftIds.has(key)}
-                    <pre class="draft-preview">{p.content}</pre>
-                  {/if}
-                </li>
-              {/each}
-            </ul>
-          </DraftCard>
-        {/snippet}
-
-        {#snippet sourceDraftCardBlock(draft: ConversationSourceDraft)}
-          <DraftCard
-            headline={`📚 ${draft.sources.length} source${draft.sources.length === 1 ? '' : 's'}`}
-            note={draft.note}
-            approveLabel="Approve & ingest"
-            onApprove={() => handleApproveSource(tab.id, draft)}
-            onDiscard={() => handleDiscardSource(tab.id, draft.draftId)}
-          >
-            <ul class="source-list">
-              {#each draft.sources as s, si (si)}
-                <li>
-                  <span class="source-kind">{sourceKindLabel(s)}</span>
-                  <span class="source-value">{sourceLabel(s)}</span>
-                </li>
-              {/each}
-            </ul>
-          </DraftCard>
-        {/snippet}
-
-        {#snippet sourceResultLine(_draftId: string, outcomes: SourceIngestOutcome[])}
-          <!-- Compact "Filed:" line that replaces the propose_sources
-               card after Approve. Each successfully filed source title
-               is a clickable link that opens the source in the editor;
-               failures are shown muted in-line so the user can see what
-               didn't land without losing the click-to-open story. The
-               line is persistent (no dismiss) — it lives in the
-               transcript so the user can scroll back later. -->
-          <div class="filed-line">
-            <span class="filed-prefix">📚 Filed:</span>
-            {#each outcomes as o, oi (oi)}
-              {#if oi > 0}<span class="filed-sep">·</span>{/if}
-              {#if o.error}
-                <span class="filed-error" title={o.error}>⚠ {sourceLabel(o.input)}</span>
-              {:else if o.sourceId}
-                <button
-                  type="button"
-                  class="filed-link"
-                  title={o.duplicate ? 'Already in library — open' : 'Open source'}
-                  onclick={() => openFiledSource(o.sourceId!)}
-                >{o.title ?? sourceLabel(o.input)}{#if o.duplicate}<span class="filed-dup"> · already in library</span>{/if}</button>
-              {:else}
-                <span class="filed-error">{sourceLabel(o.input)}</span>
-              {/if}
-            {/each}
-          </div>
-        {/snippet}
-
-        {#snippet noteResultLine(_draftId: string, filedPaths: string[])}
-          <!-- Counterpart to sourceResultLine for propose_notes. Drops
-               in where the draft card was so the user knows what filed
-               and can jump to any of the new notes with one click. -->
-          <div class="filed-line">
-            <span class="filed-prefix">📝 Filed:</span>
-            {#if filedPaths.length === 0}
-              <span class="filed-error">(no notes written)</span>
-            {:else}
-              {#each filedPaths as p, pi (p)}
-                {#if pi > 0}<span class="filed-sep">·</span>{/if}
-                <button
-                  type="button"
-                  class="filed-link"
-                  title={p}
-                  onclick={() => openFiledNote(p)}
-                >{basename(p)}</button>
-              {/each}
-            {/if}
-          </div>
-        {/snippet}
-
-        {#snippet propertyDraftCardBlock(draft: ConversationPropertyDraft)}
-          <!-- set_properties review card. Mirrors the source/note card
-               chrome but shows the proposed frontmatter patch per note
-               instead of a flat list. Each value renders with its key
-               so the user can eyeball the diff without clicking
-               through to the file. -->
-          <DraftCard
-            headline={`🔑 ${draft.updates.length} note${draft.updates.length === 1 ? '' : 's'}`}
-            note={draft.note}
-            approveLabel="Approve & apply"
-            onApprove={() => handleApproveProperty(tab.id, draft)}
-            onDiscard={() => handleDiscardProperty(tab.id, draft.draftId)}
-          >
-            <ul class="property-update-list">
-              {#each draft.updates as u, ui (ui)}
-                <li class="property-update">
-                  <div class="property-update-path">{u.relativePath}</div>
-                  <ul class="property-kv-list">
-                    {#each Object.entries(u.properties) as [k, v] (k)}
-                      <li class="property-kv" class:property-kv-delete={v === null}>
-                        <span class="property-key">{k}:</span>
-                        <span class="property-value">{formatPropertyValue(v)}</span>
-                      </li>
-                    {/each}
-                  </ul>
-                </li>
-              {/each}
-            </ul>
-          </DraftCard>
-        {/snippet}
-
-        {#snippet sourcePropertyDraftCardBlock(draft: ConversationSourcePropertyDraft)}
-          <!-- propose_source_properties review card (#103). Shows the proposed
-               abstract / TL;DR for one source; Approve upserts dc:abstract /
-               thought:tldr into its meta.ttl. -->
-          <DraftCard
-            headline="📄 Source summary"
-            note={draft.note}
-            approveLabel="Approve & apply"
-            onApprove={() => handleApproveSourceProperty(tab.id, draft)}
-            onDiscard={() => handleDiscardSourceProperty(tab.id, draft.draftId)}
-          >
-            <div class="property-update-path">{draft.sourceId}</div>
-            {#if draft.abstract}
-              <div class="source-prop-block">
-                <div class="source-prop-label">Abstract</div>
-                <div class="source-prop-text">{draft.abstract}</div>
-              </div>
-            {/if}
-            {#if draft.tldr}
-              <div class="source-prop-block">
-                <div class="source-prop-label">TL;DR</div>
-                <div class="source-prop-text">{draft.tldr}</div>
-              </div>
-            {/if}
-          </DraftCard>
-        {/snippet}
-
-        {#snippet sourcePropertyResultLine(_draftId: string, outcome: SourcePropertyOutcome)}
-          <div class="filed-line">
-            <span class="filed-prefix">📄 Updated:</span>
-            {#if outcome.error}
-              <span class="filed-error" title={outcome.error}>⚠ {outcome.sourceId}</span>
-            {:else if outcome.changedPredicates.length === 0}
-              <span class="filed-error">{outcome.sourceId} · no change</span>
-            {:else}
-              <span class="filed-link" title={outcome.sourceId}>{outcome.sourceId} · {outcome.changedPredicates.join(', ')}</span>
-            {/if}
-          </div>
-        {/snippet}
-
-        {#snippet claimsDraftCardBlock(draft: ConversationClaimsDraft)}
-          <!-- propose_claims review card (#104). Each claim shows its kind,
-               confidence, and the supporting quote; Approve files claim notes +
-               excerpt nodes through the approval engine. -->
-          <DraftCard
-            headline={`🧩 ${draft.claims.length} claim${draft.claims.length === 1 ? '' : 's'}`}
-            note={draft.note}
-            approveLabel="Approve & file"
-            onApprove={() => handleApproveClaims(tab.id, draft)}
-            onDiscard={() => handleDiscardClaims(tab.id, draft.draftId)}
-          >
-            <ul class="claims-list">
-              {#each draft.claims as c, ci (ci)}
-                <li class="claim-item">
-                  <div class="claim-head">
-                    <span class="claim-kind">{c.kind}</span>
-                    <span class="claim-conf">conf {c.confidence.toFixed(2)}</span>
-                    {#if !c.quoteFound}<span class="claim-approx" title="Quote wasn't a verbatim substring of the body — excerpt files without a character anchor">approx</span>{/if}
-                  </div>
-                  <div class="claim-text">{c.text}</div>
-                  <div class="claim-quote">{c.quote}</div>
-                </li>
-              {/each}
-            </ul>
-          </DraftCard>
-        {/snippet}
-
-        {#snippet claimsResultLine(_draftId: string, outcome: ClaimsOutcome)}
-          <div class="filed-line">
-            <span class="filed-prefix">🧩 Filed:</span>
-            {#if outcome.error}
-              <span class="filed-error" title={outcome.error}>⚠ {outcome.sourceId}</span>
-            {:else}
-              {#each outcome.claimPaths as p, pi (p)}
-                {#if pi > 0}<span class="filed-sep">·</span>{/if}
-                <button type="button" class="filed-link" title={p} onclick={() => openFiledNote(p)}>{basename(p)}</button>
-              {/each}
-              <span class="filed-dup"> · {outcome.excerptIds.length} excerpt{outcome.excerptIds.length === 1 ? '' : 's'}</span>
-            {/if}
-          </div>
-        {/snippet}
-
-        {#snippet propertyResultLine(_draftId: string, outcomes: PropertyUpdateOutcome[])}
-          <!-- Compact "Updated:" line that replaces the propose-property
-               card after Approve. Each successfully-patched path is a
-               clickable link that opens the note in the editor; the
-               count of changed keys is shown in parentheses so the user
-               can confirm the patch landed without re-reading the
-               frontmatter. -->
-          <div class="filed-line">
-            <span class="filed-prefix">🔑 Updated:</span>
-            {#if outcomes.length === 0}
-              <span class="filed-error">(no notes touched)</span>
-            {:else}
-              {#each outcomes as o, oi (oi)}
-                {#if oi > 0}<span class="filed-sep">·</span>{/if}
-                {#if o.error}
-                  <span class="filed-error" title={o.error}>⚠ {basename(o.relativePath)}</span>
-                {:else if o.changedKeys.length === 0}
-                  <button
-                    type="button"
-                    class="filed-link"
-                    title="{o.relativePath} — already up to date"
-                    onclick={() => openFiledNote(o.relativePath)}
-                  >{basename(o.relativePath)}<span class="filed-dup"> · no-op</span></button>
-                {:else}
-                  <button
-                    type="button"
-                    class="filed-link"
-                    title="{o.relativePath} — {o.changedKeys.join(', ')}"
-                    onclick={() => openFiledNote(o.relativePath)}
-                  >{basename(o.relativePath)}<span class="filed-dup"> · {o.changedKeys.length} key{o.changedKeys.length === 1 ? '' : 's'}</span></button>
-                {/if}
-              {/each}
-            {/if}
-          </div>
-        {/snippet}
-
         <div class="messages" bind:this={scrollEl}>
           <!-- Interleave: message → cards anchored to that message →
                next message. Keeps a draft visually attached to the
@@ -914,74 +415,7 @@
                card. -->
           {#each tab.conversation.messages as msg, i}
             {@render messageBlock(msg, tab, i)}
-            {#each draftsAt(tab, i) as draft (draft.draftId)}
-              {@render noteDraftCard(draft)}
-            {/each}
-            {#each sourceDraftsAt(tab, i) as draft (draft.draftId)}
-              {@render sourceDraftCardBlock(draft)}
-            {/each}
-            {#each sourceResultsAt(tab, i) as [draftId, entry] (draftId)}
-              {@render sourceResultLine(draftId, entry.outcomes)}
-            {/each}
-            {#each noteResultsAt(tab, i) as [draftId, entry] (draftId)}
-              {@render noteResultLine(draftId, entry.filedPaths)}
-            {/each}
-            {#each propertyDraftsAt(tab, i) as draft (draft.draftId)}
-              {@render propertyDraftCardBlock(draft)}
-            {/each}
-            {#each propertyResultsAt(tab, i) as [draftId, entry] (draftId)}
-              {@render propertyResultLine(draftId, entry.outcomes)}
-            {/each}
-            {#each sourcePropertyDraftsAt(tab, i) as draft (draft.draftId)}
-              {@render sourcePropertyDraftCardBlock(draft)}
-            {/each}
-            {#each sourcePropertyResultsAt(tab, i) as [draftId, entry] (draftId)}
-              {@render sourcePropertyResultLine(draftId, entry.outcome)}
-            {/each}
-            {#each claimsDraftsAt(tab, i) as draft (draft.draftId)}
-              {@render claimsDraftCardBlock(draft)}
-            {/each}
-            {#each claimsResultsAt(tab, i) as [draftId, entry] (draftId)}
-              {@render claimsResultLine(draftId, entry.outcome)}
-            {/each}
-            {#each computeDraftsAt(tab, i) as draft (draft.draftId)}
-              <ComputeDraftCard
-                {draft}
-                runState={tab.computeDraftState[draft.draftId]}
-                onRun={onRunCompute}
-                onInsert={onInsertCompute}
-                onDiscard={onDiscardCompute}
-                onOpenInserted={openInsertedNote}
-              />
-            {/each}
-            {#each refactorDraftsAt(tab, i) as draft (draft.draftId)}
-              <RefactorDraftCard
-                {draft}
-                onApprove={() => handleApproveRefactor(tab.id, draft)}
-                onDiscard={() => handleDiscardRefactor(tab.id, draft.draftId)}
-              />
-            {/each}
-            {#each reorgDraftsAt(tab, i) as draft (draft.draftId)}
-              <ReorgDraftCard
-                {draft}
-                onApprove={(selected) => handleApproveReorg(tab.id, draft, selected)}
-                onDiscard={() => handleDiscardReorg(tab.id, draft.draftId)}
-              />
-            {/each}
-            {#each deleteDraftsAt(tab, i) as draft (draft.draftId)}
-              <DeleteDraftCard
-                {draft}
-                onApprove={(selected) => handleApproveDelete(tab.id, draft, selected)}
-                onDiscard={() => handleDiscardDelete(tab.id, draft.draftId)}
-              />
-            {/each}
-            {#each noteBodyDraftsAt(tab, i) as draft (draft.draftId)}
-              <NoteBodyDraftCard
-                {draft}
-                onApprove={() => handleApproveNoteBody(tab.id, draft)}
-                onDiscard={() => handleDiscardNoteBody(tab.id, draft.draftId)}
-              />
-            {/each}
+            <DraftCards {tab} index={i} />
           {/each}
 
           {#if tab.streaming}
@@ -1042,74 +476,7 @@
                anchor pointing into thin air. Render at the bottom so
                the card is still visible; the inline interleaved render
                above will pick them up on reload. -->
-          {#each orphanDrafts(tab) as draft (draft.draftId)}
-            {@render noteDraftCard(draft)}
-          {/each}
-          {#each orphanSourceDrafts(tab) as draft (draft.draftId)}
-            {@render sourceDraftCardBlock(draft)}
-          {/each}
-          {#each orphanSourceResults(tab) as [draftId, entry] (draftId)}
-            {@render sourceResultLine(draftId, entry.outcomes)}
-          {/each}
-          {#each orphanNoteResults(tab) as [draftId, entry] (draftId)}
-            {@render noteResultLine(draftId, entry.filedPaths)}
-          {/each}
-          {#each orphanPropertyDrafts(tab) as draft (draft.draftId)}
-            {@render propertyDraftCardBlock(draft)}
-          {/each}
-          {#each orphanPropertyResults(tab) as [draftId, entry] (draftId)}
-            {@render propertyResultLine(draftId, entry.outcomes)}
-          {/each}
-          {#each orphanSourcePropertyDrafts(tab) as draft (draft.draftId)}
-            {@render sourcePropertyDraftCardBlock(draft)}
-          {/each}
-          {#each orphanSourcePropertyResults(tab) as [draftId, entry] (draftId)}
-            {@render sourcePropertyResultLine(draftId, entry.outcome)}
-          {/each}
-          {#each orphanClaimsDrafts(tab) as draft (draft.draftId)}
-            {@render claimsDraftCardBlock(draft)}
-          {/each}
-          {#each orphanClaimsResults(tab) as [draftId, entry] (draftId)}
-            {@render claimsResultLine(draftId, entry.outcome)}
-          {/each}
-          {#each orphanComputeDrafts(tab) as draft (draft.draftId)}
-            <ComputeDraftCard
-              {draft}
-              runState={tab.computeDraftState[draft.draftId]}
-              onRun={onRunCompute}
-              onInsert={onInsertCompute}
-              onDiscard={onDiscardCompute}
-              onOpenInserted={openInsertedNote}
-            />
-          {/each}
-          {#each orphanRefactorDrafts(tab) as draft (draft.draftId)}
-            <RefactorDraftCard
-              {draft}
-              onApprove={() => handleApproveRefactor(tab.id, draft)}
-              onDiscard={() => handleDiscardRefactor(tab.id, draft.draftId)}
-            />
-          {/each}
-          {#each orphanReorgDrafts(tab) as draft (draft.draftId)}
-            <ReorgDraftCard
-              {draft}
-              onApprove={(selected) => handleApproveReorg(tab.id, draft, selected)}
-              onDiscard={() => handleDiscardReorg(tab.id, draft.draftId)}
-            />
-          {/each}
-          {#each orphanDeleteDrafts(tab) as draft (draft.draftId)}
-            <DeleteDraftCard
-              {draft}
-              onApprove={(selected) => handleApproveDelete(tab.id, draft, selected)}
-              onDiscard={() => handleDiscardDelete(tab.id, draft.draftId)}
-            />
-          {/each}
-          {#each orphanNoteBodyDrafts(tab) as draft (draft.draftId)}
-            <NoteBodyDraftCard
-              {draft}
-              onApprove={() => handleApproveNoteBody(tab.id, draft)}
-              onDiscard={() => handleDiscardNoteBody(tab.id, draft.draftId)}
-            />
-          {/each}
+          <DraftCards {tab} index={null} />
         </div>
 
         <Composer {tab} {currentNotePath} {onInvokeSkill} bind:this={composer} />
@@ -1456,221 +823,6 @@
     flex-direction: column;
     gap: 8px;
   }
-  .drafts-label { font-size: 11px; color: var(--text-muted); }
-  /* Proposal / draft card (§9.2) — accent-tinted bordered card so a
-     proposal pops out of the message stream as something the user is
-     about to decide on. Uses oklch color-mix so it tracks palette swaps. */
-  /* Draft-card chrome (.draft-card / .draft-note / .draft-actions / .draft-btn)
-     now lives in DraftCard.svelte and ComputeDraftCard.svelte. Only the
-     card-body styles below (.draft-paths, source/property/claims lists) remain
-     here — they render as DraftCard children, so they keep the panel's scope. */
-  .draft-paths { list-style: none; padding: 0; margin: 0; display: flex; flex-direction: column; gap: 4px; }
-  .draft-path-btn {
-    width: 100%;
-    text-align: left;
-    background: none;
-    border: none;
-    padding: 4px 6px;
-    color: var(--text);
-    font: inherit;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    border-radius: 3px;
-  }
-  .draft-path-btn:hover { background: var(--bg, var(--bg-sidebar)); }
-  .draft-path { font-family: var(--font-mono, monospace); font-size: 12px; }
-  .draft-toggle { color: var(--text-muted); margin-left: auto; }
-  .draft-preview {
-    margin: 4px 0 4px 18px;
-    padding: 8px 10px;
-    background: var(--bg);
-    border: 1px solid var(--border);
-    border-radius: 4px;
-    font-size: 12px;
-    white-space: pre-wrap;
-    overflow-x: auto;
-    max-height: 280px;
-  }
-
-  /* propose_sources cards. Same outer chrome as note draft cards;
-     interior shows a flat list of url/identifier pills rather than the
-     expandable preview tree (we don't fetch metadata until Approve, so
-     there's nothing to preview pre-ingest). */
-  .source-list { list-style: none; padding: 0; margin: 0; display: flex; flex-direction: column; gap: 3px; }
-  .source-list li {
-    display: flex;
-    align-items: baseline;
-    gap: 8px;
-    padding: 3px 6px;
-    font-family: var(--font-mono, monospace);
-    font-size: 12px;
-  }
-  .source-kind {
-    display: inline-block;
-    min-width: 38px;
-    padding: 1px 5px;
-    border: 1px solid var(--border);
-    border-radius: 3px;
-    background: var(--bg, var(--bg-sidebar));
-    color: var(--text-muted);
-    font-size: 10px;
-    text-transform: uppercase;
-    text-align: center;
-    letter-spacing: 0.04em;
-    flex-shrink: 0;
-  }
-  .source-value {
-    color: var(--text);
-    overflow-wrap: anywhere;
-  }
-
-  /* set_properties review card interior. Each per-note patch is a
-     small block with the relative path as a header and a key:value
-     list underneath. Deleted keys render dimmed with a strikethrough
-     marker so removals are visually distinct from sets. */
-  .property-update-list {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    flex-direction: column;
-    gap: 6px;
-  }
-  .property-update {
-    border-left: 2px solid var(--border);
-    padding-left: 8px;
-  }
-  .property-update-path {
-    font-family: var(--font-mono, monospace);
-    font-size: 12px;
-    color: var(--text-muted);
-    margin-bottom: 2px;
-  }
-  /* propose_source_properties card (#103). */
-  .source-prop-block {
-    margin: 6px 0;
-    border-left: 2px solid var(--border);
-    padding-left: 8px;
-  }
-  .source-prop-label {
-    font-size: 10px;
-    text-transform: uppercase;
-    letter-spacing: 0.5px;
-    color: var(--text-muted);
-    margin-bottom: 2px;
-  }
-  .source-prop-text {
-    font-size: 12px;
-    line-height: 1.5;
-    color: var(--text);
-    white-space: pre-wrap;
-  }
-  /* propose_claims card (#104). */
-  .claims-list {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-  }
-  .claim-item {
-    border-left: 2px solid var(--border);
-    padding-left: 8px;
-  }
-  .claim-head {
-    display: flex;
-    align-items: baseline;
-    gap: 8px;
-    font-size: 10px;
-    margin-bottom: 2px;
-  }
-  .claim-kind {
-    text-transform: uppercase;
-    letter-spacing: 0.5px;
-    color: var(--text-muted);
-  }
-  .claim-conf { color: var(--text-muted); font-variant-numeric: tabular-nums; }
-  .claim-approx {
-    color: var(--bg);
-    background: var(--text-muted);
-    border-radius: 3px;
-    padding: 0 4px;
-  }
-  .claim-text { font-size: 13px; color: var(--text); }
-  .claim-quote {
-    font-size: 11px;
-    color: var(--text-muted);
-    border-left: 2px solid var(--border);
-    padding-left: 6px;
-    margin-top: 2px;
-    white-space: pre-wrap;
-  }
-  .property-kv-list {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    flex-direction: column;
-    gap: 1px;
-  }
-  .property-kv {
-    display: flex;
-    gap: 6px;
-    font-size: 12px;
-    padding: 1px 4px;
-    font-family: var(--font-mono, monospace);
-  }
-  .property-key {
-    color: var(--text-muted);
-    flex-shrink: 0;
-  }
-  .property-value {
-    color: var(--text);
-    overflow-wrap: anywhere;
-  }
-  .property-kv-delete .property-key,
-  .property-kv-delete .property-value {
-    color: var(--text-muted);
-    text-decoration: line-through;
-    text-decoration-color: color-mix(in srgb, var(--text-muted) 60%, transparent);
-  }
-
-  /* Post-Approve summary line — replaces the inline draft card for
-     both propose_notes and propose_sources. Single-line, no chrome,
-     clickable filenames/titles. Persistent (no dismiss) so the user
-     can scroll back later and still navigate to filed resources. */
-  .filed-line {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: baseline;
-    gap: 6px;
-    padding: 6px 4px;
-    font-size: 12px;
-    color: var(--text-muted);
-  }
-  .filed-prefix {
-    color: var(--text-muted);
-    flex-shrink: 0;
-  }
-  .filed-sep { color: var(--border); }
-  .filed-link {
-    background: none;
-    border: none;
-    padding: 0;
-    margin: 0;
-    color: var(--accent);
-    font: inherit;
-    cursor: pointer;
-    text-decoration: underline;
-    text-underline-offset: 2px;
-    text-decoration-color: color-mix(in srgb, var(--accent) 40%, transparent);
-  }
-  .filed-link:hover { text-decoration-color: var(--accent); }
-  .filed-dup { color: var(--text-muted); text-decoration: none; }
-  .filed-error { color: var(--text-muted); }
 
   .empty {
     flex: 1;
